[reafactoring] delete mainTest.java and move methods to KiumTradingSystemTest.java
