--- src/test/java/mainTest.java
@@ -1,99 +1,0 @@
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class mainTest {
-
-    @Test
-    void kiumloginpass() {
-        String id = "ABC";
-        String password = "BTS";
-
-        KiumTradingSystem app = new KiumTradingSystem();
-        String actual = app.login(id, password);
-        String expected = String.format("%s님 로그인 성공", id);
-        assertEquals(actual, expected);
-    }
-
-    @Test
-    void kiumloginfail() {
-        String id = "ABC";
-        String password = "TEST";
-
-        KiumTradingSystem app = new KiumTradingSystem();
-        String actual = app.login(id, password);
-        String expected = String.format("%s님 로그인 실패", id);
-        assertEquals(actual, expected);
-    }
-
-    @Test
-    void nemologinpass() {
-        String id = "ABC";
-        String password = "BTS";
-
-        NamuTradingSystem app = new NamuTradingSystem();
-        String actual = app.login(id, password);
-        String expected = String.format("%s님 로그인 성공", id);
-
-        assertEquals(actual, expected);
-    }
-
-    @Test
-    void nemologinfail() {
-        String id = "ABC";
-        String password = "TEST";
-
-        NamuTradingSystem app = new NamuTradingSystem();
-        String actual = app.login(id, password);
-        String expected = String.format("%s님 로그인 실패", id);
-        assertEquals(actual, expected);
-    }
-
-
-    void kiumbuypass() {
-        String stockCode = "T01";
-        int count = 1;
-        int price = 1000;
-
-        KiumTradingSystem app = new KiumTradingSystem();
-        String actual = app.buy(stockCode, count, price);
-        String expected = String.format("%s를 %d 가격에 매수하였음", stockCode, price);
-        assertEquals(actual, expected);
-    }
-
-    void namubuypass() {
-        String stockCode = "T01";
-        int count = 1;
-        int price = 1000;
-
-        NamuTradingSystem app = new NamuTradingSystem();
-        String actual = app.buy(stockCode, count, price);
-        String expected = String.format("%s를 %d 가격에 매수하였음", stockCode, price);
-        assertEquals(actual, expected);
-    }
-
-
-    void kiumsellpass() {
-        String stockCode = "T02";
-        int count = 1;
-        int price = 2000;
-
-        KiumTradingSystem app = new KiumTradingSystem();
-        String actual = app.buy(stockCode,count,price);
-        String expected = String.format("%s를 %d 가격에 매도하였음", stockCode,price);
-        assertEquals(actual, expected);
-    }
-
-    void namusellypass() {
-        String stockCode = "T02";
-        int count = 1;
-        int price = 2000;
-
-        NamuTradingSystem app = new NamuTradingSystem();
-        String actual = app.buy(stockCode,count,price);
-        String expected = String.format("%s를 %d 가격에 매도하였음", stockCode,price);
-        assertEquals(actual, expected);
-    }
-
-}
